% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEnrichmentMap.R
\name{getEnrichmentMap}
\alias{getEnrichmentMap}
\title{Get Enrichment Map}
\usage{
getEnrichmentMap(genes, curCluster = 1, curPathway = "RAF activation")
}
\arguments{
\item{genes}{a vector of genes}

\item{curCluster}{integer, the current cluster}

\item{curPathway}{character, the current pathway}
}
\value{
a list of outputs
}
\description{
Get Enrichment Map
}
