% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{getMatrix}
\alias{getMatrix}
\title{Matrix}
\usage{
getMatrix(n, r, c)
}
\arguments{
\item{n}{entries}

\item{r}{number of rows}

\item{c}{number of columns}
}
\description{
Matrix
}
\concept{
enrichmentmap
}
